Add the needed files

+++ likes/models.py
@@ -3,9 +3,9 @@ from django.contrib.auth.models import User
 from django.contrib.contenttypes.models import ContentType
 from django.contrib.contenttypes.fields import GenericForeignKey
 
+
 class LikedItem(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
-    #Using generic RelationShip
     content_type = models.ForeignKey(ContentType, on_delete=models.CASCADE)
-    object_id = models.PositiveSmallIntegerField()
+    object_id = models.PositiveIntegerField()
     content_object = GenericForeignKey()
